Rename scrollable_sheet_extent_scope
